+++ e2e/wallets/eth-asset-context/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @coinmasters/eth-asset-context-test
 
+## 1.0.15
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.13.18
+
 ## 1.0.14
 
 ### Patch Changes